Mon, Jan 2, 2023, 1:50 PM +00:00

+++ readability/readability.c
@@ -19,7 +19,7 @@ int main(void)
     words = word_count(input);
     sentences = sentence_count(input);
 
-    float L = (letters / words) * 100;
+    float L = ( (float)letters / words) * 100;
     float S = (sentences / words) * 100;
 
     index = round(0.0588 * L - 0.296 * S - 15.8);
